+++ hierarchical_sales_forecasting/b2b_revenue_forecasting/__init__.py
@@ -8,7 +8,7 @@ from b2b_revenue_forecasting.pipeline_adjuster import PipelineAdjuster
 from b2b_revenue_forecasting.metric_spec import MetricSpec
 from b2b_revenue_forecasting.batch import cascade_many
 
-__version__ = "0.10.0"
+__version__ = "0.10.1"
 
 __all__ = [
     "SalesHierarchy",